# packages/contracts/openapi/nova-file-api.openapi.json
{
  "components": {
    "responses": {
      "FileForbiddenResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "Canonical forbidden request response."
      },
...
      "FileInvalidRequestResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "Canonical invalid-request response."
      },
      "FileQueueUnavailableResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "Canonical queue unavailable response."
      },
      "FileUnauthorizedResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "Canonical unauthorized request response."
      }
    },
    "schemas": {
...
      "EnqueueJobRequest": {
        "additionalProperties": false,
        "description": "Request payload for job enqueue endpoint.",
        "properties": {
          "job_type": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Job Type",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "job_type"
        ],
        "title": "EnqueueJobRequest",
        "type": "object"
      },
      "EnqueueJobResponse": {
        "additionalProperties": false,
        "description": "Response payload for enqueue endpoint.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "EnqueueJobResponse",
        "type": "object"
      },
...
      "JobListResponse": {
        "additionalProperties": false,
        "description": "Response payload for job listing endpoint.",
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobRecord"
            },
            "maxItems": 200,
            "title": "Jobs",
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
...
      "JobResultUpdateRequest": {
        "additionalProperties": false,
        "allOf": [
          {
            "if": {
              "properties": {
                "status": {
                  "const": "succeeded"
                }
              },
              "required": [
                "status"
              ]
            },
            "then": {
              "properties": {
                "error": {
                  "type": "null"
                }
              }
            }
          }
        ],
        "description": "Worker/internal request payload for job result updates.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "JobResultUpdateRequest",
        "type": "object"
      },
      "JobResultUpdateResponse": {
        "additionalProperties": false,
        "description": "Response payload for internal job result updates.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "updated_at"
        ],
        "title": "JobResultUpdateResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "X-Worker-Token": {
        "description": "Worker token header for trusted job-worker calls.",
        "in": "header",
        "name": "X-Worker-Token",
        "type": "apiKey"
      },
      "sessionAuth": {
        "description": "Session identifier header for caller context.",
        "in": "header",
        "name": "X-Session-Id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "nova-file-api",
    "version": "0.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/internal/jobs/{job_id}/result": {
      "post": {
        "description": "Update job status/result from trusted worker-side processing.",
        "operationId": "update_job_result",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Worker-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobResultUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "X-Worker-Token": []
          }
        ],
        "summary": "Update Job Result",
        "tags": [
          "jobs"
        ],
        "x-nova-sdk-visibility": "internal"
      }
    },
    "/v1/jobs": {
      "get": {
        "description": "List caller-owned jobs with most recent first.",
        "operationId": "list_jobs",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Enqueue async processing job and return job id.",
        "operationId": "create_job",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          },
          "503": {
            "$ref": "#/components/responses/FileQueueUnavailableResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "description": "Return status for the caller-owned job.",
        "operationId": "get_job_status",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Get Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "Cancel a caller-owned non-terminal job.",
        "operationId": "cancel_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "description": "Return poll events with an SSE-compatible envelope.",
        "operationId": "list_job_events",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "List Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/retry": {
      "post": {
        "description": "Retry a terminal failed or canceled job.",
        "operationId": "retry_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Retry Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/transfers/downloads/presign": {
      "post": {
        "description": "Issue presigned GET URL for caller-scoped key.",
        "operationId": "presign_download",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignDownloadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Presign Download",
        "tags": [
          "transfers"
        ]
      }
    },
    "/v1/transfers/uploads/abort": {
      "post": {
        "description": "Abort multipart upload.",
        "operationId": "abort_upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbortUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Abort Upload",
        "tags": [
          "transfers"
        ]
      }
    },
    "/v1/transfers/uploads/complete": {
      "post": {
        "description": "Complete multipart upload.",
        "operationId": "complete_upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Complete Upload",
        "tags": [
          "transfers"
        ]
      }
    },
    "/v1/transfers/uploads/initiate": {
      "post": {
        "description": "Choose upload strategy and return presigned metadata.",
        "operationId": "initiate_upload",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "409": {
            "$ref": "#/components/responses/FileIdempotencyConflictResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          },
          "503": {
            "$ref": "#/components/responses/FileIdempotencyUnavailableResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Initiate Upload",
        "tags": [
          "transfers"
        ]
      }
    },
    "/v1/transfers/uploads/sign-parts": {
      "post": {
        "description": "Return presigned multipart part URLs.",
        "operationId": "sign_upload_parts",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignPartsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignPartsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/FileUnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/FileForbiddenResponse"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "summary": "Sign Upload Parts",
        "tags": [
          "transfers"
        ]
      }
    }
  }
}